Power per weapon, plus Thor's optimisations (measured)
WEAPON POWER. In the original every mounted weapon carries its own power
with its own cap (weapons.md sec.1, field +0x34 / +0x24); collecting the
same weapon again raises it. Until now a re-collect only set an already set
bit and did nothing at all. Damage, the beam stage and the beam's
(power+1)x3 now follow the WEAPON'S power, not the ship's - the ship stage
stays with the default forward shot, which is not in lvl_weapon_* and has
no power of its own. The POWER article and the POWER token raise every
mounted weapon, and the article stays in stock while any weapon can still
grow rather than while the ship stage can. The power belongs to the
loadout, so it goes into the checkpoint snapshot.

Deliberate deviation, documented at the table: in the original a re-collect
of a SIDE weapon claims another free position instead of raising the power.
That needs a mount to own a weapon rather than the other way round. As long
as one weapon has at most one position, the alternative would be a pickup
that visibly does nothing.

THOR'S OPTIMISATIONS, two fixed scenes, same frame count on both sides:
  deterministic enemies + containers  143479 -> 141202 cycles/frame  -1.6%
  normal play                         123362 -> 121464 cycles/frame  -1.5%
Taken: SetSprite/SetSpriteEx/SetSpritePosition parameters u8 -> u16 (a u8
parameter is zero-extended twice per call), the two OAM words written
without the byte-splitting detour, rects_overlap widths likewise, and
spr_draw_s_single_flip folded into one SetSpriteEx.

fps_spd_s loses its MUL and DIV - but NOT the way it was sent. The negative
branch was (m-1) >> 1, and MEASURED in this toolchain the signed division
truncates towards zero while the shift rounds down: at m = -6 that is -3
against -4, so every leftward and upward speed would have come out a pixel
too large at 20 fps. The shift now only ever sees a non-negative value.
Proven equal to the old formula over all 1201 values from -600 to +600.

Not taken: the rest of the archive is an older base (SHIP_CELLS 7,
OAM_POOL_BASE 16, the beam without its reference masking, build_bar_assets
without bar_pal_load) and would have undone today's fixes.

+++ library.h
@@ -200,11 +200,11 @@ void PrintHex2(u8 Plane, u8 PaletteNo, u8 x, u8 y, u32 Value, u8 Len);
 void PrintString(u8 Plane, u8 Palette, u8 XPos, u8 YPos, const char * theString);
 
 extern u16 g_oam_pos_shadow[64];   /* position write cache, see SetSpritePosition */
-void SetSprite(u8 SpriteNo, u16 TileNo, u8 Chain, u8 XPos, u8 YPos, u8 PaletteNo);
+void SetSprite(u16 SpriteNo, u16 TileNo, u16 Chain, u16 XPos, u16 YPos, u16 PaletteNo);
 /* SetSprite plus SpriteControl in one pass, Ctrl = Priority | Flips. Saves a
    far call and an OAM read-modify-write per sprite - see library.c. */
-void SetSpriteEx(u8 SpriteNo, u16 TileNo, u8 Chain, u8 XPos, u8 YPos, u8 PaletteNo, u8 Ctrl);
-void SetSpritePosition(u8 SpriteNo, u8 XPos, u8 YPos);
+void SetSpriteEx(u16 SpriteNo, u16 TileNo, u16 Chain, u16 XPos, u16 YPos, u16 PaletteNo, u16 Ctrl);
+void SetSpritePosition(u16 SpriteNo, u16 XPos, u16 YPos);
 
 s32 Multiply32bit(s32 Value1, s32 Value2);
 void SeedRandom(void);
